+++ src/explore-education-statistics-admin/src/pages/release/data/components/DataFilesTable.tsx
@@ -15,6 +15,7 @@ import {
 import {
   DataFile,
   DataFileImportStatus,
+  DeleteDataFilePlan,
 } from '@admin/services/releaseDataFileService';
 import ButtonGroup from '@common/components/ButtonGroup';
 import ButtonText from '@common/components/ButtonText';
@@ -22,6 +23,7 @@ import Modal from '@common/components/Modal';
 import React from 'react';
 import { generatePath } from 'react-router';
 import styles from './DataFilesTable.module.scss';
+import DataFilesTableRow from './DataFilesTableRow';
 
 interface Props {
   canUpdateRelease?: boolean;
@@ -30,7 +32,7 @@ interface Props {
   publicationId: string;
   releaseVersionId: string;
   testId?: string;
-  onDeleteFile: (dataFile: DataFile) => Promise<void>;
+  onDeleteFile: (deletedFileId: string) => void;
   onStatusChange: (
     dataFile: DataFile,
     importStatus: DataFileImportStatus,
@@ -62,144 +64,15 @@ export default function DataFilesTable({
 
       <tbody>
         {dataFiles.map(dataFile => (
-          <tr key={dataFile.title}>
-            <td data-testid="Title" className={styles.title}>
-              {dataFile.title}
-            </td>
-            <td data-testid="Size" className={styles.fileSize}>
-              {dataFile.fileSize.size.toLocaleString()} {dataFile.fileSize.unit}
-            </td>
-            <td data-testid="Status">
-              <ImporterStatus
-                className={styles.fileStatus}
-                dataFile={dataFile}
-                hideErrors
-                releaseVersionId={releaseVersionId}
-                onStatusChange={onStatusChange}
-              />
-            </td>
-            <td data-testid="Actions">
-              <ButtonGroup className={styles.actions}>
-                <Modal
-                  showClose
-                  title="Data file details"
-                  triggerButton={<ButtonText>View details</ButtonText>}
-                >
-                  <DataFileSummaryList
-                    dataFile={dataFile}
-                    releaseVersionId={releaseVersionId}
-                    onStatusChange={onStatusChange}
-                  />
-                </Modal>
-                {canUpdateRelease &&
-                  terminalImportStatuses.includes(dataFile.status) && (
-                    <>
-                      {dataFile.status === 'COMPLETE' && (
-                        <>
-                          <Link
-                            to={generatePath<ReleaseDataFileRouteParams>(
-                              releaseDataFileRoute.path,
-                              {
-                                publicationId,
-                                releaseVersionId,
-                                fileId: dataFile.id,
-                              },
-                            )}
-                          >
-                            Edit title
-                          </Link>
-                          {dataFile.publicApiDataSetId ? (
-                            <Modal
-                              showClose
-                              title="Cannot replace data"
-                              triggerButton={
-                                <ButtonText>Replace data</ButtonText>
-                              }
-                            >
-                              <p>
-                                This data file has an API data set linked to it.
-                                Please remove the API data set before replacing
-                                the data.
-                              </p>
-                              <p>
-                                <Link
-                                  to={generatePath<ReleaseDataSetRouteParams>(
-                                    releaseApiDataSetDetailsRoute.path,
-                                    {
-                                      publicationId,
-                                      releaseVersionId,
-                                      dataSetId: dataFile.publicApiDataSetId,
-                                    },
-                                  )}
-                                >
-                                  Go to API data set
-                                </Link>
-                              </p>
-                            </Modal>
-                          ) : (
-                            <Link
-                              to={generatePath<ReleaseDataFileReplaceRouteParams>(
-                                releaseDataFileReplaceRoute.path,
-                                {
-                                  publicationId,
-                                  releaseVersionId,
-                                  fileId: dataFile.id,
-                                },
-                              )}
-                            >
-                              Replace data
-                            </Link>
-                          )}
-                        </>
-                      )}
-                      {dataFile.publicApiDataSetId ? (
-                        <Modal
-                          showClose
-                          title="Cannot delete files"
-                          triggerButton={
-                            <ButtonText variant="warning">
-                              Delete files
-                            </ButtonText>
-                          }
-                        >
-                          <p>
-                            This data file has an API data set linked to it.
-                            Please remove the API data set before deleting.
-                          </p>
-                          <p>
-                            <Link
-                              to={generatePath<ReleaseDataSetRouteParams>(
-                                releaseApiDataSetDetailsRoute.path,
-                                {
-                                  publicationId,
-                                  releaseVersionId,
-                                  dataSetId: dataFile.publicApiDataSetId,
-                                },
-                              )}
-                            >
-                              Go to API data set
-                            </Link>
-                          </p>
-                        </Modal>
-                      ) : (
-                        <ButtonText
-                          onClick={() => onDeleteFile(dataFile)}
-                          variant="warning"
-                        >
-                          Delete files
-                        </ButtonText>
-                      )}
-                    </>
-                  )}
-                {dataFile.permissions.canCancelImport && (
-                  <DataUploadCancelButton
-                    releaseVersionId={releaseVersionId}
-                    fileId={dataFile.id}
-                  />
-                )}
-              </ButtonGroup>
-            </td>
-          </tr>
+          <DataFilesTableRow
+            canUpdateRelease={canUpdateRelease}
+            dataFile={dataFile}
+            key={dataFile.title}
+            publicationId={publicationId}
+            releaseVersionId={releaseVersionId}
+            onConfirmDelete={onDeleteFile}
+            onStatusChange={onStatusChange}
+          />
         ))}
       </tbody>
     </table>
